.vscode-server/data/User/History/114ac8d2/ddT7.py
import jwt
import requests
from django.http import JsonResponse
from django.utils.decorators import method_decorator
from django.views import View
from django.views.decorators.csrf import csrf_exempt


#from accountapp.models import AppUser

