.vscode-server/data/User/History/3b85e74c/rfs7.py
from allauth.account.views import LogoutView
from accounts.views import index

urlpatterns = [

    path('accounts/', include('allauth.urls')),
    path('accounts/logout/', LogoutView.as_view(), name='account_logout'),
    
]