requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
attrs==20.3.0
aws-cfn-bootstrap==2.0
awscli==2.15.30
awscrt==0.19.19
Babel==2.9.1
backports.tarfile==1.2.0
cffi==1.14.5
chardet==4.0.0
check-wheel-contents==0.6.0
chevron==0.13.1
click==8.1.7
cloud-init==22.2.2
colorama==0.4.4
configobj==5.0.6
cryptography==36.0.1
dbus-python==1.2.18
distro==1.5.0
dj-rest-auth==6.0.0
Django==4.2.16
django-allauth==64.2.1
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docutils==0.21.2
drf-spectacular==0.27.2
drf-yasg==1.21.7
ec2-hibinit-agent==1.0.8
environ==1.0
gpg==1.15.1
idna==2.10
importlib_metadata==8.5.0
inflection==0.5.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
jeepney==0.8.0
Jinja2==2.11.3
jmespath==0.10.0
jsonpatch==1.21
jsonpointer==2.0
jsonschema==3.2.0
keyring==25.4.1
libcomps==0.1.20
lockfile==0.12.2
markdown-it-py==3.0.0
MarkupSafe==1.1.1
mdurl==0.1.2
more-itertools==10.5.0
netifaces==0.10.6
nh3==0.2.18
oauthlib==3.0.2
packaging==22.0
pkginfo==1.10.0
ply==3.11
prettytable==0.7.2
prompt-toolkit==3.0.24
pycparser==2.20
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
PyMySQL==1.1.1
pyparsing==2.4.7
pyrsistent==0.17.3
pyserial==3.4
PySocks==1.7.1
python-daemon==2.3.0
python-dateutil==2.8.1
pytz==2022.7.1
PyYAML==5.4.1
readme_renderer==44.0
release-notification==1.2
requests==2.25.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.8.1
rpm==4.16.1.3
ruamel.yaml==0.16.6
ruamel.yaml.clib==0.1.2
SecretStorage==3.3.3
selinux==3.4
sepolicy==3.4
setools==4.4.1
six==1.15.0
sqlparse==0.5.1
support-info==1.0
systemd-python==235
tomli==2.0.1
twine==5.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.20
uWSGI==2.0.26
wcwidth==0.2.5
wheel-filename==1.4.1
zipp==3.20.2
